+++ server/api/v1/controllers/index.js
@@ -1,0 +1,5 @@
+import Auth from './Auth';
+import Ride from './Ride';
+import RideRequest from './RideRequest';
+
+export { Auth, Ride, RideRequest };
