+++ server/api/v1/middlewares/jwtverify.js
@@ -1,5 +1,6 @@
 import jwt from 'jsonwebtoken';
 
+
 const jwtverify = (req, res, next) => {
   let tokengen = null;
   if (req.headers.authorization && req.headers.authorization.split(' ')[0] === 'Bearer') {
