+++ server/api/v1/routes/auth.route.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import { signup, signin, verifyUser } from '../controllers/auth.controller';
+import { Auth } from '../controllers//index';
 import jwtverify from '../middlewares/jwtverify';
 
 const route = express.Router();
 
-route.post('/auth/signup', signup);
-route.post('/auth/signin', signin);
-route.get('/auth/me', jwtverify, verifyUser);
+route.post('/auth/signup', Auth.signup);
+route.post('/auth/signin', Auth.signin);
+route.get('/auth/me', jwtverify, Auth.verifyUser);
 
 export default route;
